Add files via upload
Value stream mapping (VSM) is defined as a lean tool that employs a flowchart documenting every step in the process. Many lean practitioners see VSM as a fundamental tool to identify waste, reduce process cycle times, and implement process improvement. This is about the production in coco-cola factory , what are  there materials , how much is produced , how much is wasted , how much is preserved. The calculation is done in excel sheet while the other two files are about the value stream map diagram calculating the cycle time and log time.